Use pybind for rd_file
This fixes a bug with the report_steps and num_report_steps
methods, which would throw KeyError on files without "SEQNUM".
It now returns []/0 as documented.

The `has_kw` and `report_dates` functions had imprecise docstrings,
which were updated. The module docstring referenced rd_file.c which
no longer exists: so this internal detail was removed from the
docstring.

+++ lib/resdata/rd_grid.cpp
@@ -2616,7 +2616,7 @@ static int rd_grid_dual_porosity_GRID_check(rd::File *rd_file) {
     else {
         int dualp_flag = FILEHEAD_DUAL_POROSITY;
         size_t num_corners = rd_file->num_named_kw(CORNERS_KW);
-        int matrix_index = 0;
+        size_t matrix_index = 0;
         size_t fracture_index;
 
         rd_kw_type *matrix_kw;
